Broke out various components in our ReportsCenter into separate files
