+++ src/views/ReportsCenter/ReportsCenter.styl
@@ -10,6 +10,12 @@
         box-shadow: none;
     }
 
+    #leave-moderator-note {
+        textarea#moderator-note {
+            width: 100%;
+        }
+    }
+
     #move-tree-container {
         height: 2rem;
     }
