+++ src/views/ReportsCenter/UserHistory.tsx
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C)  Online-Go.com
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import * as React from "react";
+import { Player } from "Player";
+import { ModTools } from "User";
+
+export function UserHistory({ user }: { user: any }): JSX.Element {
+    if (!user) {
+        return null;
+    }
+
+    return (
+        <>
+            <h3>
+                History for <Player user={user} />
+            </h3>
+            <ModTools user_id={user.id} show_mod_log={true} collapse_same_users={true} />
+        </>
+    );
+}
